requirements.txt
fastapi~=0.115.5
uvicorn
langchain~=0.3.7
chromadb~=0.5.18
openai
pydantic~=2.9.2
python-dotenv~=1.0.1
pytest
httpx
langchain-community~=0.3.7
torch
sentence_transformers
langchain_chroma
langchain_openai
langchain-openai~=0.2.8
pydantic-settings~=2.6.1